refact: redesign video sink design
Remove utilwidgets.py
