+++ cv2PySide6/__init__.py
@@ -1,19 +1,16 @@
 from .version import __version__ # noqa
 
-
-from .utilwidgets import ClickableSlider
 from .labels import ScalableQLabel, NDArrayLabel
-from .videoplayer import (FrameToArrayConverter, ArrayProcessor,
-    NDArrayVideoPlayerWidget)
+from .videoplayer import (ClickableSlider, FrameToArrayConverter,
+    ArrayProcessor, NDArrayVideoPlayerWidget)
 from .util import get_data_path
 
 
 __all__ = [
-    'ClickableSlider',
-
     'ScalableQLabel',
     'NDArrayLabel',
 
+    'ClickableSlider',
     'FrameToArrayConverter',
     'ArrayProcessor',
     'NDArrayVideoPlayerWidget',
